Fix import paths for renamed directory
